+++ src/Notifications/Entities/ClientToken/ClientTokenStatus.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Paddle\SDK\Notifications\Entities\ClientToken;
+
+use Paddle\SDK\PaddleEnum;
+
+/**
+ * @method static ClientTokenStatus Active()
+ * @method static ClientTokenStatus Revoked()
+ */
+final class ClientTokenStatus extends PaddleEnum
+{
+    private const Active = 'active';
+    private const Revoked = 'revoked';
+}